Move api mocks to services.
