+++ coral/src/app/pages/index.tsx
@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { AuthUser, getAuthUser } from "src/domain/auth-user";
 import { mockUserAuthRequest } from "src/domain/auth-user/auth-user-api.msw";
 import { useEffect } from "react";
-import { getWindowWithMswInstance } from "src/domain/api-mocks/window-msw";
+import { getWindowWithMswInstance } from "src/services/api-mocks/window-msw";
 
 const HomePage = () => {
   useEffect(() => {
